gemini_computer_use/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o gemini_computer_use/requirements.txt
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via pytest
greenlet==3.2.4
    # via playwright
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
packaging==25.0
    # via pytest
playwright==1.55.0
    # via
    #   computer-use-boilerplates (pyproject.toml)
    #   pytest-playwright
pluggy==1.6.0
    # via pytest
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via
    #   pytest-base-url
    #   pytest-playwright
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-playwright==0.7.1
    # via computer-use-boilerplates (pyproject.toml)
python-dotenv==1.2.1
    # via computer-use-boilerplates (pyproject.toml)
python-slugify==8.0.4
    # via pytest-playwright
requests==2.32.5
    # via pytest-base-url
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via pyee
urllib3==2.5.0
    # via requests
